requirements.txt
streamlit
cerebras-cloud-sdk
exa-py
python-dotenv
streamlit-webrtc
openai-whisper
elevenlabs
